Move guts of examples/MultiCar/ into examples/Cars/.
