+++ drake/examples/Cars/models/breadtruck.urdf
