+++ drake/examples/Cars/models/sedan.urdf
