+++ drake/examples/Cars/trivial_car_state.h
