--- drake/examples/MultiCar/trivial_car.cc
@@ -1,1 +1,0 @@
-#include "drake/examples/MultiCar/trivial_car.h"
